Changed page names and added title to homepage
