+++ frontend/pages/map.js
@@ -1,0 +1,7 @@
+import MapComponent from "../components/map/MapComponent";
+
+const Map = () => {
+  return <MapComponent />;
+};
+
+export default Map;
